chore: move antlr classes
