--- src/main/java/com/github/lombrozo/xnav/AntlrChardata.java
@@ -1,73 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2025 Volodya Lombrozo
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package com.github.lombrozo.xnav;
-
-import java.util.Optional;
-import java.util.stream.Stream;
-import org.w3c.dom.Node;
-
-public final class AntlrChardata implements Xml {
-
-    private final XMLParser.ChardataContext context;
-
-    public AntlrChardata(final XMLParser.ChardataContext context) {
-        this.context = context;
-    }
-
-    @Override
-    public Xml child(final String element) {
-        throw new UnsupportedOperationException("Not implemented");
-    }
-
-    @Override
-    public Optional<Xml> attribute(final String name) {
-        throw new UnsupportedOperationException("Not implemented");
-    }
-
-    @Override
-    public Optional<String> text() {
-        return Optional.ofNullable(this.context.getText());
-    }
-
-    @Override
-    public Stream<Xml> children() {
-        return Stream.empty();
-    }
-
-    @Override
-    public String name() {
-        return "";
-    }
-
-    @Override
-    public Xml copy() {
-        throw new UnsupportedOperationException("Not implemented");
-    }
-
-    @Override
-    public Node node() {
-        throw new UnsupportedOperationException("Not implemented");
-    }
-}
