+++ src/test/java/com/github/lombrozo/xnav/AntlrAttribute.java
@@ -29,79 +29,54 @@ import java.util.stream.Stream;
 import lombok.EqualsAndHashCode;
 import org.w3c.dom.Node;
 
-/**
- * element
- *     : '<' Name attribute* '>' content '<' '/' Name '>'
- *     | '<' Name attribute* '/>'
- *     ;
- */
 @EqualsAndHashCode
-public final class AntlrXmlElement implements Xml {
+final class AntlrAttribute implements Xml {
 
     @EqualsAndHashCode.Exclude
-    XMLParser.ElementContext context;
+    private final XMLParser.AttributeContext context;
 
-    public AntlrXmlElement(XMLParser.ElementContext context) {
+    public AntlrAttribute(final XMLParser.AttributeContext context) {
         this.context = context;
     }
 
     @Override
     public Xml child(final String element) {
-        final XMLParser.ContentContext content = this.context.content();
-        for (final XMLParser.ElementContext context : content.element()) {
-            final AntlrXmlElement elem = new AntlrXmlElement(context);
-            if (elem.name().equals(element)) {
-                return elem;
-            }
-        }
-        return new Empty();
+        return null;
     }
 
     @Override
     public Optional<Xml> attribute(final String name) {
-        return this.context.attribute()
-            .stream()
-            .map(attr -> (Xml) new AntlrAttribute(attr))
-            .filter(attr -> attr.name().equals(name))
-            .findFirst();
-//        if (this.context.attribute() == null) {
-//            return Optional.empty();
-//        } else {
-//            return Optional.empty();
-//        }
-//        return new AntlrXmlAttributes(this.context.attribute());
-    }
-
-    @Override
-    public Stream<Xml> children() {
-        if (this.context.content() == null) {
-            return Stream.empty();
-        }
-        return new AntlrXmlContent(this.context.content()).children();
+        return Optional.empty();
     }
 
     @Override
     public Optional<String> text() {
-        return new AntlrXmlContent(this.context.content()).text();
+        final String text = this.context.STRING().getText();
+        return Optional.of(text.substring(1, text.length() - 1));
+    }
+
+    @Override
+    public Stream<Xml> children() {
+        return null;
     }
 
     @Override
     public String name() {
-        return this.context.Name(0).getText();
+        return this.context.Name().getText();
     }
 
     @Override
     public Xml copy() {
-        return new AntlrXmlElement(this.context);
+        return null;
     }
 
     @Override
     public Node node() {
-        throw new UnsupportedOperationException("Not implemented");
+        return null;
     }
 
-    @Override
     @EqualsAndHashCode.Include
+    @Override
     public String toString() {
         return this.context.getText();
     }
