+++ src/test/java/com/github/lombrozo/xnav/AntlrXmlDocument.java
@@ -41,7 +41,7 @@ final class AntlrXmlDocument implements Xml {
 
     private final String xml;
 
-    public AntlrXmlDocument(final String... xml) {
+    AntlrXmlDocument(final String... xml) {
         this(String.join("", xml));
     }
